+++ CHANGELOG.md
@@ -11,7 +11,9 @@ The most recent changes are on top, in each type of changes category.
 
 - [Bugfix] places input should keep state [#427](https://github.com/cyfronet-fid/sat4envi/issues/427)
 
+### Added
 
+- Notifications module [#423](https://github.com/cyfronet-fid/sat4envi/issues/423)
 
 ## [v5.1.0]